+++ tests/api_test.py
@@ -7769,9 +7769,12 @@ class InputSavedVJPTest(jtu.JaxTestCase):
     def f(x, y):
       return x * y
 
-    primals = 2., 3.
-    y, f_vjp = api.si_vjp(f, [True, True], *primals)
-    arg_cts = f_vjp(1., *primals)
+    primals = [2., 3.]
+    y, f_vjp = jax.vjp(f, *primals)
+    f_vjp.args_res = [None, None]
+    y_grad = 1.
+    f_vjp.args_res = primals
+    arg_cts = f_vjp(1.)
     self.assertAllClose(y, 6.)
     self.assertAllClose(arg_cts, (3., 2.))
 
@@ -7782,28 +7785,19 @@ class InputSavedVJPTest(jtu.JaxTestCase):
     @jax.jit
     def g():
       primals = 2., 3.
-      y, f_vjp = api.si_vjp(f, [True, True], *primals)
+      y, f_vjp = jax.vjp(f, *primals)
+      f_vjp.args_res = [None, None]
       return y, f_vjp
 
     @jax.jit
     def h(f_vjp):
-      return f_vjp(1., 2., 3.)
+      f_vjp.args_res = [2., 3.]
+      return f_vjp(1.)
 
     y, f_vjp = g()
     arg_cts = h(f_vjp)
     self.assertAllClose(y, 6.)
     self.assertAllClose(arg_cts, (3., 2.))
-
-  def test_basic_unused(self):
-    f = jnp.sin
-    primals = 3.,
-    y, f_vjp = api.si_vjp(f, [True], *primals)
-    x_ct, = f_vjp(1., *primals)
-    self.assertAllClose(y, jnp.sin(3.))
-    self.assertAllClose(x_ct, jnp.cos(3.))
-
-    with self.assertRaisesRegex(Exception, "not used by the backward pass: x"):
-      _ = api.si_vjp(f, [True], *primals, allow_unused=False)
 
   def test_basic_unused_vjp3(self):
     f = jnp.sin
@@ -7814,58 +7808,28 @@ class InputSavedVJPTest(jtu.JaxTestCase):
     self.assertAllClose(x_ct, jnp.cos(3.))
     self.assertIsInstance(f_vjp.args_res[0], api.NotNeeded)  # can check if unused
 
-  def test_basic_opaque(self):
-    f = jnp.sin
-    primals = 3.,
-    with self.assertRaisesRegex(Exception, "the backward pass requires opaque"):
-      _ = api.si_vjp(f, [True], *primals, allow_opaque=False)
-
   def test_basic_opaque_vjp3(self):
     f = jnp.sin
     primals = 3.,
     _, f_vjp = api.vjp(f, *primals)
-    assert f_vjp.opaque_residuals  # can detect if opaque res are used
+    self.assertTrue(f_vjp.opaque_residuals)  # can detect if opaque res are used
 
   def test_basic_pytree_error(self):
     def f(x):
       return [x['hi'] * x['bye']]
 
-    y, f_vjp = api.si_vjp(f, [True], {'hi': 2., 'bye': 3.})
-    arg_ct, = f_vjp([1.], {'hi': 2., 'bye': 3.})
+    y, f_vjp = jax.vjp(f, {'hi': 2., 'bye': 3.})
+    f_vjp.args_res = [None]
+    y_grad = [1.]
+    f_vjp.args_res = [{'hi': 2., 'bye': 3.}]
+    arg_ct, = f_vjp(y_grad)
     self.assertAllClose(y, [6.])
     self.assertAllClose(arg_ct, {'hi': 3., 'bye': 2.})
 
-    with self.assertRaisesRegex(ValueError, "but the structures differ"):
-      f_vjp(1., {'hi': 2.})
-
-  # TODO(mattjj): improve this vjp3 error message
-  # def test_basic_pytree_error_vjp3(self):
-  #   def f(x):
-  #     return [x['hi'] * x['bye']]
-
-  #   y, f_vjp = api.vjp(f, {'hi': 2., 'bye': 3.})
-  #   arg_ct, = f_vjp([1.], {'hi': 2., 'bye': 3.})
-  #   self.assertAllClose(y, [6.])
-  #   self.assertAllClose(arg_ct, {'hi': 3., 'bye': 2.})
-
-  #   f_vjp.args_res[0] = {'hi': 2.}
-  #   with self.assertRaisesRegex(ValueError, "but the structures differ"):
-  #     f_vjp(1.)
-
-  def test_fsdp(self):
-    # see https://github.com/jax-ml/jax/pull/27017 for why this is called "fsdp"
-    def f2(x, w):
-      x = 1. * x
-      x = x @ w
-      x = 2. * x
-      return x
-
-    x = jnp.ones((3, 4))
-    w = jnp.ones((4, 4))
-    y, f2_sivjp = api.si_vjp(f2, [False, True], x, w)
-    y_grad = jnp.ones_like(y)
-    x_grad, w_grad = f2_sivjp(y_grad, w)
-    self.assertAllClose(x_grad, 2. * y_grad @ w.T)
+    # TODO(mattjj): Raise an error message.
+    # with self.assertRaisesRegex(ValueError, "but the structures differ"):
+    #   f_vjp.args_res = [{'hi': 2.}]
+    #   f_vjp([1.])
 
   def test_fsdp_error(self):
     # see https://github.com/jax-ml/jax/pull/27017 for why this is called "fsdp"
@@ -7877,10 +7841,12 @@ class InputSavedVJPTest(jtu.JaxTestCase):
 
     x = jnp.ones((3, 4))
     w = jnp.ones((4, 4))
-    y, f2_sivjp = api.si_vjp(f2, [False, True], x, w)
+    y, f2_vjp = jax.vjp(f2, x, w)
+    f2_vjp.args_res[1] = None
     y_grad = jnp.ones((2, 4))
+    f2_vjp.args_res[1] = w
     with self.assertRaisesRegex(ValueError, "unexpected JAX type"):
-      f2_sivjp(y_grad, w)
+      f2_vjp(y_grad)
 
   def test_fsdp_vjp3(self):
     # see https://github.com/jax-ml/jax/pull/27017 for why this is called "fsdp"
@@ -7902,9 +7868,10 @@ class InputSavedVJPTest(jtu.JaxTestCase):
     self.assertAllClose(w_grad, 2. * x.T @ y_grad)
 
   def test_doesnt_leak_symbolic_zeros(self):
-    _, vjp = api.si_vjp(lambda x: 1., [False], 3.14)
+    _, vjp = jax.vjp(lambda x: 1., 3.14)
     ans, = vjp(1.0)
     self.assertIsInstance(ans, jax.Array)
+
 
 class TracebackTest(jtu.JaxTestCase):
   # These tests are to catch regressions in Python traceback sizes. Our